Fold UpdateReleaseNotes code into ReleaseManager
