--- tools/Google.Cloud.Tools.UpdateReleaseNotes/Google.Cloud.Tools.UpdateReleaseNotes.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <IsPackable>false</IsPackable>
-    <Description>Console app to update release note documentation from git logs</Description>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="LibGit2Sharp" Version="0.25.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Google.Cloud.Tools.Common\Google.Cloud.Tools.Common.csproj" />
-  </ItemGroup>
-</Project>
