+++ tools/Google.Cloud.Tools.sln
@@ -30,8 +30,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Cloud.Tools.SourceMa
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Cloud.Tools.CheckVersionCompatibility", "Google.Cloud.Tools.CheckVersionCompatibility\Google.Cloud.Tools.CheckVersionCompatibility.csproj", "{416C9A5E-B09B-42BC-A6B3-886308671E61}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Cloud.Tools.UpdateReleaseNotes", "Google.Cloud.Tools.UpdateReleaseNotes\Google.Cloud.Tools.UpdateReleaseNotes.csproj", "{E66FC669-1F15-4158-8980-4958565F84EE}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Cloud.Tools.CompareVersions", "Google.Cloud.Tools.CompareVersions\Google.Cloud.Tools.CompareVersions.csproj", "{BC2FA087-5154-49BF-AEE1-5B50305BAAAD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Cloud.Tools.ReleaseManager", "Google.Cloud.Tools.ReleaseManager\Google.Cloud.Tools.ReleaseManager.csproj", "{4171E471-ABAB-43B9-B759-6641BA6C4A4A}"
@@ -216,18 +214,6 @@ Global
 		{416C9A5E-B09B-42BC-A6B3-886308671E61}.Release|x64.Build.0 = Release|Any CPU
 		{416C9A5E-B09B-42BC-A6B3-886308671E61}.Release|x86.ActiveCfg = Release|Any CPU
 		{416C9A5E-B09B-42BC-A6B3-886308671E61}.Release|x86.Build.0 = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|x64.Build.0 = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Debug|x86.Build.0 = Debug|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|x64.ActiveCfg = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|x64.Build.0 = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|x86.ActiveCfg = Release|Any CPU
-		{E66FC669-1F15-4158-8980-4958565F84EE}.Release|x86.Build.0 = Release|Any CPU
 		{BC2FA087-5154-49BF-AEE1-5B50305BAAAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BC2FA087-5154-49BF-AEE1-5B50305BAAAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BC2FA087-5154-49BF-AEE1-5B50305BAAAD}.Debug|x64.ActiveCfg = Debug|Any CPU
